+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/plugin/logic/channel/internal/ChannelLogic.kt
@@ -398,7 +398,7 @@ internal class ChannelLogic(
      * @param message [Message].
      */
     internal fun incrementUnreadCountIfNecessary(message: Message) {
-        val currentUserId = GlobalMutableState.getOrCreate().user.value?.id ?: return
+        val currentUserId = globalMutableState.user.value?.id ?: return
 
         val shouldIncrementUnreadCount =
             message.shouldIncrementUnreadCount(